Problem: No Bit / Byte / String parsecomb in Lean
Solution:

 - Start the simplest possible universal parser combinator in Straume.Zeptoparsec

The idea of this is to be able to trivially parse bits, bytes or strings!
This will make sure that we can actually work with every single data type that's interesting to be parsed and that discrete electronic devices stream.

We should then be able to attach any Zeptoparsec parser to a particular instance of a Stream and make it chunk up stuff.