Problem: take needs new data, but Uni doesn't contain it
Solution:

It's ok, it's by design. All that Uni (and any other Straume structure) contains is just an elaborate context of reading from the underlying data source using functions like takeN etc.
To model the idea of producing data, we came up with MonadEmit, which is something that says that a source can emit a value. Now we can use this as a constrint in our "methods". Cool stuff.