Generalise Chunk into Terminable
Problem: we have the Chunk inductive type, which models finality
info tacked onto values. But this is in principle just one way to
express finality, and we want to keep Straume as general as possible.

Solution: defined Terminable, a type class that models the same as
Chunk, but, like, more generally, man.