Envision Aka class; add chunkLength
Problem: there's no way to consume infinite streams atom by atom or
chunk by chunk. Consequently, Avots.lean needs heavy re-writing with
better models we have come up with.
Also, the original API includes chunkLength; we don't have it yet.

Solution:
The solution builds on top of Iterable. We observe that even though the
streams are infinite (countably), or effectively infinite
(larger-than-RAM) we need to have a way to bet chunks. Iterable is
something finite and something that allows us to go from a composite
type to an atomic type. We call values of the composite type "chunks"
and values of the atomic type "atoms".

The other component of infinite stream modelling is a minimalist
MonadEmit, which provides a generic interface to get something out of
an infinite stream.

Given all that, we can now make a several typeclasses which together
provide both capacity to extract things from an infinite stream
atom-by-atom (for parser implementation) and chunk-by-chunk
(for parser plumbing). Note that this approach is more generic than
Stream from Megaparsec because these typeclasses together grant a lot
of freedom to define any way to view streams.

So, we have renamed Avots.lean to Aka.lean (/source/ to /well/),
left in only a minimal Aka type class, as we want to add all necessary
functions outside of it and restrict them with various type classes.
Added chunkLength function.