Add example of Zeptoparsec usage to Main.lean
Problem: we now have Straume which can stream large files 🎉, but
by itself it's quite useless for further data processing.
However, we do have Zeptoparsec that can help us! We just need an
example of a marginally less trivial parser in Main.lean.

Solution: we consider the following basic (but not trivial!) example
of counting the number of sentences shorter than 50 characters from
the same Chapter 1 of Pride and Prejudice. "Sentence" is defined as
separated by a dot, so not actually a sentence in many cases, but
it's fine. Our job here isn't to write a full "lexer" for sentences,
but to show the approach.
Added this thusly motivated example to Main.lean.